Move show logic to App
